+++ release-notes.md
@@ -12,9 +12,9 @@
 * breaking change: "sprintly-uikit" ---> "sprintly-ui" and UMD global rename
 * breaking change: new clearer directory structure
 * breaking change: adds gulpfile and new convenience scripts to package.json
-* adds dev and test server support
 * adds coverage reporting via istanbul for browser tests
 * better wercker build integration and coveralls.io integration
 * updates dependencies
-* condense examples to single index.html file
+* adds dev and test server support
+* add index.html file to examples for testing new components during dev
 * updates README